requirements.txt
streamlit
pandas
scikit-learn==1.4.0
joblib
matplotlib
seaborn
nltk
numpy
flask
flask-cors
langchain
langchain-google-genai
python-dotenv
pydantic